requirements.txt
pycryptodome
httpx
cryptography